Add recipe_items sync and probably a bunch of other stuff I forgot

+++ datasync/src/main.rs
@@ -1,4 +1,4 @@
-use std::{error::Error, fs::File, io::BufReader, path::Path};
+use std::path::Path;
 
 use git2::Repository;
 use simple_logger::SimpleLogger;
@@ -9,39 +9,65 @@ mod types;
 const REPOSITORY_URL: &str = "https://github.com/xivapi/ffxiv-datamining";
 
 fn main() {
-    SimpleLogger::from_env().init().unwrap();
+    SimpleLogger::from_env()
+        .with_module_level("tokio_postgres", log::LevelFilter::Warn)
+        .init()
+        .unwrap();
 
     let mut datamining_path = std::env::temp_dir();
     datamining_path.push("ffxiv-datamining");
 
     log::info!("Cloning repository {}", REPOSITORY_URL);
-    // clone_repository(&datamining_path);
+    clone_repository(&datamining_path);
 
     let mut client = db::Client::new(
         "host=localhost port=1231 user=postgres password=postgres dbname=craftup_dev",
     )
     .unwrap();
 
-    let mut items_path = datamining_path.clone();
-    items_path.push("csv");
-    items_path.push("Item");
-    items_path.set_extension("csv");
-    let items = read_data_file::<types::Item>(&items_path).unwrap();
-    &client.add_items(items).unwrap();
-
-    let mut recipes_path = datamining_path.clone();
-    recipes_path.push("csv");
-    recipes_path.push("Recipe");
-    recipes_path.set_extension("csv");
-    let recipes = read_data_file::<types::Recipe>(&recipes_path).unwrap();
-    &client.add_recipes(recipes).unwrap();
+    add_entity::<types::Item>(&mut client, "Item", db::Client::add_items).unwrap();
+    add_entity::<types::RecipeLevel>(
+        &mut client,
+        "RecipeLevelTable",
+        db::Client::add_recipe_levels,
+    )
+    .unwrap();
+    add_entity::<types::Recipe>(&mut client, "Recipe", db::Client::add_recipes).unwrap();
 
     log::info!("Done! Bye!");
 }
 
-fn read_data_file<T>(file_source: &Path) -> Result<Vec<T>, Box<dyn Error>>
+fn add_entity<T>(
+    client: &mut db::Client,
+    entity_name: &str,
+    f: fn(&mut db::Client, Vec<T>) -> Result<(), Box<dyn std::error::Error>>,
+) -> Result<(), Box<dyn std::error::Error>>
 where
     T: for<'de> serde::Deserialize<'de>,
+    T: types::GameItem,
+{
+    let mut datamining_path = std::env::temp_dir();
+    datamining_path.push("ffxiv-datamining");
+
+    let mut entity_path = datamining_path.clone();
+    entity_path.push("csv");
+    entity_path.push(entity_name);
+    entity_path.set_extension("csv");
+
+    log::debug!("Parsing {}", entity_path.to_str().unwrap());
+    let entities = read_data_file::<T>(&entity_path)?;
+
+    log::info!("Synchronizing {} {}s", entities.len(), entity_name);
+    f(client, entities)?;
+
+    log::info!("{}s synchronized.", entity_name);
+    Ok(())
+}
+
+fn read_data_file<T>(file_source: &Path) -> Result<Vec<T>, Box<dyn std::error::Error>>
+where
+    T: for<'de> serde::Deserialize<'de>,
+    T: types::GameItem,
 {
     let contents: &str = &std::fs::read_to_string(file_source)
         .unwrap()
@@ -54,15 +80,18 @@ where
         .has_headers(false)
         .from_reader(contents.as_bytes())
         .deserialize()
-        .map(|x| x.unwrap())
+        .map(|x| -> T { x.unwrap() })
+        .filter(|x| x.is_valid())
         .collect::<Vec<T>>();
 
     Ok(results)
 }
 
 fn clone_repository(destination: &Path) {
-    log::info!("Removing path {}", destination.to_str().unwrap());
-    std::fs::remove_dir_all(&destination).unwrap();
+    if destination.exists() {
+        log::info!("Removing path {}", destination.to_str().unwrap());
+        std::fs::remove_dir_all(&destination).unwrap();
+    }
 
     log::info!("Creating path {}", destination.to_str().unwrap());
     std::fs::create_dir(&destination).unwrap();
